+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+unreleased
+-----------------------
+* Task #22993 - Add a tab to ReportekEngine management where we can see migrations
+
 3.7.2-dev11 (10-02-2015)
 -----------------------